Type the literals, and stop deleting the compiler's explanation
Three rounds of "ambiguous use of <some SwiftUI modifier>" in files with no
Mapbox in them, each naming a different modifier. They are one bug.

A ternary of bare numeric literals — `pulse ? 1.12 : 0.86` — leaves the
literal's type for the solver to choose, and several overloads of
scaleEffect, opacity, offset, frame and shadow will accept it. It stays
latent until the module's overload set is crowded enough to make the choice
genuinely undecidable, which is why the Mapbox build hits it and CI's
fallback build does not, and why it lands in a file that never imports
Mapbox. Every one of the fifteen sites in TimiNowUI is now wrapped in
CGFloat(...) or Double(...), so the whole class is gone rather than the two
lines that happened to be reported this time.

The reason it took three rounds is mine. summarise_failure grepped for
"error:" and printed nothing else — and Swift explains an ambiguity in the
`note: found this candidate` lines directly beneath, which is precisely what
that grep discarded. So the diagnosis the compiler had already written was
thrown away, and I guessed at it instead. It now prints each error with its
notes, joining each note to the candidate signature under it so the pair
survives de-duplication.

Guarded: no ternary of bare numeric literals may be passed to an overloaded
SwiftUI modifier in TimiNowUI. Verified by reintroducing three of them.

+++ apps/customer-mobile/Sources/TimiNowUI/OfferAndTrackerViews.swift
@@ -71,7 +71,7 @@ struct OfferSearchView: View {
             Picker("Sort offers", selection: $sort) { Text("Recommended").tag("recommended"); Text("Closest").tag("distance"); Text("Shortest wait").tag("wait"); Text("Lowest deposit").tag("cost") }.pickerStyle(.segmented)
             ForEach(Array(offers.enumerated()), id: \.element.id) { index, offer in
                 OfferCard(offer: offer, rank: index + 1, isWorking: store.isWorking) { Task { await store.selectOffer(offer) } }
-                    .offset(y: appeared.contains(offer.id) ? 0 : 22).opacity(appeared.contains(offer.id) ? 1 : 0)
+                    .offset(y: CGFloat(appeared.contains(offer.id) ? 0 : 22)).opacity(Double(appeared.contains(offer.id) ? 1 : 0))
                     .onAppear { withAnimation(.spring(response: 0.48, dampingFraction: 0.82).delay(Double(index) * 0.08)) { _ = appeared.insert(offer.id) } }
             }
             Text("Availability and waits are reported by clinics and may change. Emergency hospitals independently triage every arriving patient.").font(.caption).foregroundStyle(TimiColor.muted).padding(.top, 6)
